Backported to 1.21.10

+++ src/main/java/fr/madu59/ptp/ProjectileInfo.java
@@ -234,10 +234,10 @@ public class ProjectileInfo {
 
         Vec3 pos = new Vec3(player.getX(), player.getEyeY() - 0.30000001192092896, player.getZ());
 
-        float g = Mth.sin((double)(player.getXRot() * 0.017453292F));
-        float h = Mth.cos((double)(player.getXRot() * 0.017453292F));
-        float i = Mth.sin((double)(player.getYRot() * 0.017453292F));
-        float j = Mth.cos((double)(player.getYRot() * 0.017453292F));
+        float g = Mth.sin(player.getXRot() * 0.017453292F);
+        float h = Mth.cos(player.getXRot() * 0.017453292F);
+        float i = Mth.sin(player.getYRot() * 0.017453292F);
+        float j = Mth.cos(player.getYRot() * 0.017453292F);
         float k = 0.5F * 6.2831855F;
         float l = 0.02F * 0.5F;
         Vec3 vel = new Vec3((double)(-i * h * 0.3F) + Math.cos((double)k) * (double)l, (double)(-g * 0.3F + 0.1F), (double)(j * h * 0.3F) + Math.sin((double)k) * (double)l);
@@ -256,9 +256,9 @@ public class ProjectileInfo {
     }
 
     private static Vec3 AngleFromRot(float f, float g, float h){
-        float k = -Mth.sin((double)(g * 0.017453292F)) * Mth.cos((double)(f * 0.017453292F));
-        float l = -Mth.sin((double)((f + h) * 0.017453292F));
-        float m = Mth.cos((double)(g * 0.017453292F)) * Mth.cos((double)(f * 0.017453292F));
+        float k = -Mth.sin(g * 0.017453292F) * Mth.cos(f * 0.017453292F);
+        float l = -Mth.sin((f + h) * 0.017453292F);
+        float m = Mth.cos(g * 0.017453292F) * Mth.cos(f * 0.017453292F);
 
         return new Vec3((double)k, (double)l, (double)m).normalize();
     }
